refactor: 인수테스트 공통 설정하는 클래스 추가
